jet/features/rag_web_search_chatbot/requirements.txt
langchain
langchain-ollama
langchain-community
langchain-text-splitters
langgraph
chromadb
tavily-python
streamlit
python-dotenv
pytest
